requirements.txt
beautifulsoup4==4.9.3
boto3==1.17.39
botocore==1.20.39
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
flask-googlemaps==0.4.1
Flask-PyMongo==2.3.0
future==0.18.2
geographiclib==1.50
glob2==0.7
haversine==2.3.0
htmlmin==0.1.12
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.0.1
jsmin==2.2.2
livereload==2.6.3
lunr==0.5.8
Markdown==3.3.4
MarkupSafe==1.1.1
mkdocs==1.1.2
mkdocs-material==5.5.14
mkdocs-material-extensions==1.0.1
mkdocs-minify-plugin==0.3.0
nltk==3.5
numpy==1.20.2
Pygments==2.8.1
pymdown-extensions==8.1.1
pymongo==3.11.3
python-dateutil==2.8.1
python-dotenv==0.17.0
PyYAML==5.4.1
regex==2021.3.17
requests==2.25.1
reverse-geocoder==1.5.1
s3transfer==0.3.6
scipy==1.6.2
six==1.15.0
soupsieve==2.2.1
tornado==6.1
tqdm==4.59.0
urllib3==1.26.4
Werkzeug==1.0.1
